Move audio playback logic to Instrument
