+++ SwiftUI Piano Recording Task/SwiftUI Piano Recording Task/ContentView.swift
@@ -1,6 +1,8 @@
 import SwiftUI
 
 struct ContentView: View {
+    let instrument = Instrument(noteRange: 48..<72)
+
     var body: some View {
         VStack {
             Text("Piano Recorder")
@@ -10,7 +12,11 @@ struct ContentView: View {
                 .cornerRadius(8)
                 .padding(.bottom)
 
-            Piano(noteRange: 48 ..< 72)
+            Piano(noteRange: instrument.noteRange) { midiNumber in
+                instrument.playNote(midiNumber: midiNumber)
+            } onStopNote: { midiNumber in
+                instrument.stopNote(midiNumber: midiNumber)
+            }
         }
         .padding()
         .onAppear {
