+++ SwiftUI Piano Recording Task/SwiftUI Piano Recording Task/Piano.swift
@@ -4,7 +4,7 @@ import AVFoundation
 public typealias MIDINumber = Int
 public typealias NoteRange = CountableRange<MIDINumber>
 
-private extension MIDINumber {
+extension MIDINumber {
     var pitchClass: MusicalNote {
         let rawValue = self % MusicalNote.allCases.count
         let musicalNote = MusicalNote(rawValue: rawValue)
@@ -34,10 +34,8 @@ extension PianoKeyColor {
 typealias NoteCallback = ((_ midiNumber: MIDINumber) -> Void)
 
 struct Piano: View {
-    private static let KEY_FADEOUT_IN_MS = 450
     
     let notesForWhiteKeys: [MIDINumber]
-    let audioPlayers: [MIDINumber: AVAudioPlayer]
     @State var activeKeys: Set<MIDINumber> = []
 
     var onPlayNote: NoteCallback?
@@ -45,23 +43,17 @@ struct Piano: View {
     
     init(noteRange: NoteRange, onPlayNote: NoteCallback? = nil, onStopNote: NoteCallback? = nil) {
         notesForWhiteKeys = noteRange.filter { $0.pitchClass.pianoKeyColor == .white }
-        self.audioPlayers = noteRange.createAudioPlayers()
         self.onPlayNote = onPlayNote
         self.onStopNote = onStopNote
     }
     
     private func playNote(midiNumber: MIDINumber) {
         activeKeys.insert(midiNumber)
-        guard let player = audioPlayers[midiNumber] else { return }
-        player.volume = 1
-        player.currentTime = 0
-        player.play()
         onPlayNote?(midiNumber)
     }
     
     private func stopNote(midiNumber: MIDINumber) {
         activeKeys.remove(midiNumber)
-        audioPlayers[midiNumber]?.setVolume(0, fadeDuration: Double(Piano.KEY_FADEOUT_IN_MS) / 1000)
         onStopNote?(midiNumber)
     }
     
@@ -106,17 +98,4 @@ struct Piano: View {
 }
 
 
-private extension NoteRange {
-    func createAudioPlayers() -> [MIDINumber: AVAudioPlayer] {
-        var audioPlayers: [MIDINumber: AVAudioPlayer] = [:]
-        self.forEach { midiNumber in
-            let noteName = midiNumber.pitchClass.noteName.replacing("♯", with: "#")
-            let filename = "acoustic_grand_piano-mp3-2-\(noteName)\(midiNumber.octave)"
-            
-            let url = Bundle.main.url(forResource: filename, withExtension: "mp3")!
-            let player = try! AVAudioPlayer(contentsOf: url)
-            audioPlayers[midiNumber] = player
-        }
-        return audioPlayers
-    }
-}
+
